requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
requests==2.31.0
python-telegram-bot==20.5
dateparser==1.1.8
transformers>=4.0.0
torch>=1.7.0
chromadb>=0.4.0
numpy>=1.20.0
pydantic>=2.0.0
python-multipart>=0.0.5